Fix legacy mapping & zero axes
I thought we'd be able to just interleave axes, but it turns
out this is quite sensitive to non-zero biases in the 'other'
stick, so I found rotations were turning into spirals. Instead
we optimised for one stick use, where we assume someone
using legacy controls isn't using a second stick to tweak the
less-important axes.